Add postgreDict

+++ src/main/java/com/xiaoleilu/hutool/db/SqlBuilder.java
@@ -1,11 +1,14 @@
 package com.xiaoleilu.hutool.db;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map.Entry;
 
 import com.xiaoleilu.hutool.CollectionUtil;
 import com.xiaoleilu.hutool.StrUtil;
+import com.xiaoleilu.hutool.exceptions.DbRuntimeException;
 
 /**
  * SQL构建器<br>
@@ -24,6 +27,15 @@ public class SqlBuilder {
 	 */
 	public static SqlBuilder create(){
 		return new SqlBuilder();
+	}
+	
+	/**
+	 * 创建SQL构建器
+	 * @param wrapper 包装器
+	 * @return SQL构建器
+	 */
+	public static SqlBuilder create(Wrapper wrapper){
+		return new SqlBuilder(wrapper);
 	}
 	//--------------------------------------------------------------- Static methods end
 	
@@ -69,16 +81,36 @@ public class SqlBuilder {
 	}
 	//--------------------------------------------------------------- Enums end
 	
-	private StringBuilder sql = new StringBuilder();
+	final private StringBuilder sql = new StringBuilder();
 	/** 占位符对应的值列表 */
-	private List<Object> paramValues = new ArrayList<Object>();
+	final private List<Object> paramValues = new ArrayList<Object>();
+	/** 包装器 */
+	private Wrapper wrapper;
+	
+	//--------------------------------------------------------------- Constructor start
+	public SqlBuilder() {
+	}
+	
+	public SqlBuilder(Wrapper wrapper) {
+		this.wrapper = wrapper;
+	}
+	//--------------------------------------------------------------- Constructor end
 	
 	//--------------------------------------------------------------- Builder start
 	/**
 	 * 插入
 	 * @param entity 实体
+	 * @return 自己
 	 */
 	public SqlBuilder insert(Entity entity){
+		//验证
+		DbUtil.validateEntity(entity);
+		
+		if(null != wrapper) {
+			//包装字段名
+			entity = wrapper.wrap(entity);
+		}
+		
 		sql.append("INSERT INTO ").append(entity.getTableName()).append(" (");
 
 		final StringBuilder placeHolder = new StringBuilder(") VALUES (");
@@ -102,8 +134,18 @@ public class SqlBuilder {
 	/**
 	 * 删除
 	 * @param tableName 表名
+	 * @return 自己
 	 */
 	public SqlBuilder delete(String tableName){
+		if(StrUtil.isBlank(tableName)) {
+			throw new DbRuntimeException("Table name is blank !");
+		}
+		
+		if(null != wrapper) {
+			//包装表名
+			tableName = wrapper.wrap(tableName);
+		}
+		
 		sql.append("DELETE FROM ").append(tableName);
 		
 		return this;
@@ -112,8 +154,17 @@ public class SqlBuilder {
 	/**
 	 * 更新
 	 * @param entity 要更新的实体
+	 * @return 自己
 	 */
 	public SqlBuilder update(Entity entity){
+		//验证
+		DbUtil.validateEntity(entity);
+		
+		if(null != wrapper) {
+			//包装字段名
+			entity = wrapper.wrap(entity);
+		}
+		
 		sql.append("UPDATE ").append(entity.getTableName()).append(" SET ");
 		for (Entry<String, Object> entry : entity.entrySet()) {
 			if (paramValues.size() > 0) {
@@ -130,8 +181,19 @@ public class SqlBuilder {
 	 * 查询
 	 * @param isDistinct 是否添加DISTINCT关键字（查询唯一结果）
 	 * @param fields 查询的字段
+	 * @return 自己
 	 */
 	public SqlBuilder select(boolean isDistinct, String... fields){
+		return select(isDistinct, Arrays.asList(fields));
+	}
+	
+	/**
+	 * 查询
+	 * @param isDistinct 是否添加DISTINCT关键字（查询唯一结果）
+	 * @param fields 查询的字段
+	 * @return 自己
+	 */
+	public SqlBuilder select(boolean isDistinct, Collection<String> fields){
 		sql.append("SELECT ");
 		if(isDistinct) {
 			sql.append("DISTINCT ");
@@ -140,6 +202,10 @@ public class SqlBuilder {
 		if (CollectionUtil.isEmpty(fields)) {
 			sql.append("*");
 		} else {
+			if(null != wrapper) {
+				//包装字段名
+				fields = wrapper.wrap(fields);
+			}
 			sql.append(CollectionUtil.join(fields, StrUtil.COMMA));
 		}
 		
@@ -149,16 +215,36 @@ public class SqlBuilder {
 	/**
 	 * 查询（非Distinct）
 	 * @param fields 查询的字段
+	 * @return 自己
 	 */
 	public SqlBuilder select(String... fields){
 		return select(false, fields);
 	}
 	
 	/**
+	 * 查询（非Distinct）
+	 * @param fields 查询的字段
+	 * @return 自己
+	 */
+	public SqlBuilder select(Collection<String> fields){
+		return select(false, fields);
+	}
+	
+	/**
 	 * 添加 from语句
 	 * @param tableName 表名列表（多个表名用于多表查询）
+	 * @return 自己
 	 */
 	public SqlBuilder from(String... tableNames){
+		if(CollectionUtil.isEmpty(tableNames) || StrUtil.hasBlank(tableNames)) {
+			throw new DbRuntimeException("Table name is blank in table names !");
+		}
+		
+		if(null != wrapper) {
+			//包装表名
+			tableNames = wrapper.wrap(tableNames);
+		}
+		
 		sql.append(" FROM ").append(CollectionUtil.join(tableNames, StrUtil.COMMA));
 		
 		return this;
@@ -169,19 +255,35 @@ public class SqlBuilder {
 	 * 只支持单一的逻辑运算符（例如多个条件之间）
 	 * @param logicalOperator 逻辑运算符
 	 * @param conditions 条件
+	 * @return 自己
 	 */
 	public SqlBuilder where(LogicalOperator logicalOperator, Condition... conditions){
-		sql.append(" WHERE ").append(buildCondition(logicalOperator, conditions));
+		if(CollectionUtil.isNotEmpty(conditions)) {
+			if(null != wrapper) {
+				//包装字段名
+				conditions = wrapper.wrap(conditions);
+			}
+			
+			sql.append(" WHERE ").append(buildCondition(logicalOperator, conditions));
+		}
 		
 		return this;
 	}
 	
 	/**
 	 * 分组
-	 * @param field 字段
+	 * @param fields 字段
+	 * @return 自己
 	 */
-	public SqlBuilder groupBy(String... field){
-		sql.append(" GROUP BY ").append(CollectionUtil.join(field, StrUtil.COMMA));
+	public SqlBuilder groupBy(String... fields){
+		if(CollectionUtil.isNotEmpty(fields)) {
+			if(null != wrapper) {
+				//包装字段名
+				fields = wrapper.wrap(fields);
+			}
+			
+			sql.append(" GROUP BY ").append(CollectionUtil.join(fields, StrUtil.COMMA));
+		}
 		
 		return this;
 	}
@@ -190,9 +292,17 @@ public class SqlBuilder {
 	 * 添加Having语句
 	 * @param logicalOperator 逻辑运算符
 	 * @param conditions 条件
+	 * @return 自己
 	 */
 	public SqlBuilder having(LogicalOperator logicalOperator, Condition... conditions){
-		sql.append(" HAVING ").append(buildCondition(logicalOperator, conditions));
+		if(CollectionUtil.isNotEmpty(conditions)) {
+			if(null != wrapper) {
+				//包装字段名
+				conditions = wrapper.wrap(conditions);
+			}
+			
+			sql.append(" HAVING ").append(buildCondition(logicalOperator, conditions));
+		}
 		
 		return this;
 	}
@@ -201,9 +311,18 @@ public class SqlBuilder {
 	 * 排序
 	 * @param field 按照哪个字段排序
 	 * @param order 排序方式（升序还是降序）
+	 * @return 自己
 	 */
 	public SqlBuilder orderBy(Order order, String... fields){
-		sql.append(" ORDER BY ").append(CollectionUtil.join(fields, StrUtil.COMMA)).append(StrUtil.SPACE).append(order);
+		if(CollectionUtil.isNotEmpty(fields)) {
+			if(null != wrapper) {
+				//包装字段名
+				fields = wrapper.wrap(fields);
+			}
+			
+			sql.append(" ORDER BY ").append(CollectionUtil.join(fields, StrUtil.COMMA)).append(StrUtil.SPACE)
+			.append(null == order ? StrUtil.EMPTY : order);
+		}
 		
 		return this;
 	}
@@ -212,9 +331,23 @@ public class SqlBuilder {
 	 * 多表关联
 	 * @param tableName 被关联的表名
 	 * @param join 内联方式
+	 * @return 自己
 	 */
 	public SqlBuilder join(String tableName, Join join){
-		sql.append(StrUtil.SPACE).append(join.toString()).append(" JOIN ").append(tableName);
+		if(StrUtil.isBlank(tableName)) {
+			throw new DbRuntimeException("Table name is blank !");
+		}
+		
+		if(null != join) {
+			sql.append(StrUtil.SPACE).append(join);
+		}
+		
+		if(null != wrapper) {
+			//包装表名
+			tableName = wrapper.wrap(tableName);
+		}
+		
+		sql.append(" JOIN ").append(tableName);
 		
 		return this;
 	}
@@ -224,9 +357,30 @@ public class SqlBuilder {
 	 * 只支持单一的逻辑运算符（例如多个条件之间）
 	 * @param logicalOperator 逻辑运算符
 	 * @param conditions 条件
+	 * @return 自己
 	 */
 	public SqlBuilder on(LogicalOperator logicalOperator, Condition... conditions){
-		sql.append(" ON ").append(buildCondition(logicalOperator, conditions));
+		if(CollectionUtil.isNotEmpty(conditions)) {
+			if(null != wrapper) {
+				//包装字段名
+				conditions = wrapper.wrap(conditions);
+			}
+			
+			sql.append(" ON ").append(buildCondition(logicalOperator, conditions));
+		}
+		
+		return this;
+	}
+	
+	/**
+	 * 追加SQL其它部分
+	 * @param sqlPart SQL其它部分
+	 * @return 自己
+	 */
+	public SqlBuilder append(Object sqlPart){
+		if(null != sqlPart) {
+			this.sql.append(sqlPart);
+		}
 		
 		return this;
 	}
@@ -253,7 +407,7 @@ public class SqlBuilder {
 	 * @return 构建好的SQL语句
 	 */
 	public String build(){
-		return this.sql.toString();
+		return this.sql.toString().trim();
 	}
 	
 	@Override
@@ -269,6 +423,13 @@ public class SqlBuilder {
 	 * @return 构建后的SQL语句条件部分
 	 */
 	private String buildCondition(LogicalOperator logicalOperator, Condition... conditions){
+		if(CollectionUtil.isEmpty(conditions)) {
+			return StrUtil.EMPTY;
+		}
+		if(null == logicalOperator) {
+			logicalOperator = LogicalOperator.AND;
+		}
+		
 		final StringBuilder conditionStr = new StringBuilder();
 		boolean isFirst = true;
 		for (Condition condition : conditions) {
